+++ fNbt/Tags/NbtLong.cs
@@ -51,7 +51,7 @@ namespace fNbt {
 
         #region Reading / Writing
 
-        internal override bool ReadTag(NbtBinaryReader readStream) {
+        internal override bool ReadTag(NbtBinaryReader readStream, int depthBudget) {
             if (readStream.Selector != null && !readStream.Selector(this)) {
                 readStream.ReadInt64();
                 return false;
@@ -61,12 +61,12 @@ namespace fNbt {
         }
 
 
-        internal override void SkipTag(NbtBinaryReader readStream) {
+        internal override void SkipTag(NbtBinaryReader readStream, int depthBudget) {
             readStream.ReadInt64();
         }
 
 
-        internal override void WriteTag(NbtBinaryWriter writeStream) {
+        internal override void WriteTag(NbtBinaryWriter writeStream, int depthBudget) {
             writeStream.Write(NbtTagType.Long);
             if (Name == null) throw new NbtFormatException("Name is null");
             writeStream.Write(Name);
@@ -74,7 +74,7 @@ namespace fNbt {
         }
 
 
-        internal override void WriteData(NbtBinaryWriter writeStream) {
+        internal override void WriteData(NbtBinaryWriter writeStream, int depthBudget) {
             writeStream.Write(Value);
         }
 
@@ -87,7 +87,7 @@ namespace fNbt {
         }
 
 
-        internal override void PrettyPrint(StringBuilder sb, string indentString, int indentLevel) {
+        internal override void PrettyPrint(StringBuilder sb, string indentString, int indentLevel, int depthBudget) {
             for (int i = 0; i < indentLevel; i++) {
                 sb.Append(indentString);
             }