+++ fNbt/Tags/NbtString.cs
@@ -59,7 +59,7 @@ namespace fNbt {
 
         #region Reading / Writing
 
-        internal override bool ReadTag(NbtBinaryReader readStream) {
+        internal override bool ReadTag(NbtBinaryReader readStream, int depthBudget) {
             if (readStream.Selector != null && !readStream.Selector(this)) {
                 readStream.SkipString();
                 return false;
@@ -69,12 +69,12 @@ namespace fNbt {
         }
 
 
-        internal override void SkipTag(NbtBinaryReader readStream) {
+        internal override void SkipTag(NbtBinaryReader readStream, int depthBudget) {
             readStream.SkipString();
         }
 
 
-        internal override void WriteTag(NbtBinaryWriter writeStream) {
+        internal override void WriteTag(NbtBinaryWriter writeStream, int depthBudget) {
             writeStream.Write(NbtTagType.String);
             if (Name == null) throw new NbtFormatException("Name is null");
             writeStream.Write(Name);
@@ -82,7 +82,7 @@ namespace fNbt {
         }
 
 
-        internal override void WriteData(NbtBinaryWriter writeStream) {
+        internal override void WriteData(NbtBinaryWriter writeStream, int depthBudget) {
             writeStream.Write(Value);
         }
 
@@ -95,7 +95,7 @@ namespace fNbt {
         }
 
 
-        internal override void PrettyPrint(StringBuilder sb, string indentString, int indentLevel) {
+        internal override void PrettyPrint(StringBuilder sb, string indentString, int indentLevel, int depthBudget) {
             for (int i = 0; i < indentLevel; i++) {
                 sb.Append(indentString);
             }